reorganize TF provider documentation directory
Initially, we used an open interpretation about the documentation
directory for the TF provider. It turns out that it seems to be
a moving target. They enforced improved standards on each new
applicant.

For instance, if you compare with AWS provider, they are not using
ERB helper function, we are. If you check how Random and AWS
provider organize their structures, they differ.

This commit aims to please their latest iteration of requirements,
but it is not impossible that we may have to reorganize them again.

GitOrigin-RevId: f49b38665d2cb6e27acad1fc7a985ccfceabc3de