re-work Terraform computed fields
The Terraform SDK includes two types of fields that can be computed by the backend instead of specified by the user. One is readonly fields - these can never be set by the user. The other is optional fields - fields that can be set by the user, but will be computed by the backend if not specified.

Previously, we represented optional computed fields by setting readonly = false and computed = true. This resulted in a Terraform field with Optional: true, Computed: false, and a DiffSuppressFunc. We represented readonly computed fields by setting readonly = true and computed = true on the field, resulting in a Terraform field with Computed: true.

A problem arose with port overrides. The portOverride field is not readonly, by any stretch of the imagination. But we set it to readonly = true to signal to the Terraform templates that the field should be Computed: true.

We now have a way to represent the portOverride situation. It is readonly: false, terraform_computed: 'readonly'. Readonly for Terraform, optional for everyone else.

This re-work resulted in essentially a no-op change to the generated Terraform provider, which is what we want. We just want to be able to set readonly = false on portOverride while keeping the Terraform provider exactly the same.

GitOrigin-RevId: 6ac2b7f145b0489641dc6c301beb260f4dc6e7bf