improve Terraform access rules diff suppression to tolerate more differences
When Terraform submits an access policy, the server parses and reformats it. So the next time Terraform queries it, the resource IDs might be in a different order, the formatting might be different, etc. This change makes Terraform tolerate these differences and not feel the need to update the role every time it runs.

GitOrigin-RevId: 13f92ae88ca4d09055240bee7150dc616689c823