terraform-provider-sdm: fix terraform optional computed fields
For some reason, any time we need an optional computed field in Terraform, we've been setting it to Optional: true, Computed: false, and adding a DiffSuppressFunc that checks if the value is an empty string, and if so, it suppresses any diffs.

As far as I can tell, this is just a hacky version of Optional: true, Computed: true. I'm not sure why we didn't just do that from the beginning. I think it was my fault.

GitOrigin-RevId: ca617908d60d64f4b45f0cd458d6ba24e7446a77